Final commit with .gitignore and working UI

+++ SmarterBackup.Core/Models/BackupTask.cs
@@ -14,7 +14,7 @@ namespace SmarterBackup.Core.Models
         public bool UseEncryption { set; get; }
         public string? EncryptionPassword { get; set; }
         public string? ScheduleCorn { get; set; }// هر شب ساعت 12   
-        public int? IntervalMinutes { get; set; }// مثلا 60 برای یک ساعت
+        public double? IntervalMinutes { get; set; }// مثلا 60 برای یک ساعت
 
     }
 }
